block: add __com.redhat_change-backing-file qmp command
Message-id: <c7fc7cefc7588dca2d02b0cc0051cdbdadbf20ac.1411096194.git.jcody@redhat.com>
Patchwork-id: 61313
O-Subject: [PATCH qemu-kvm-rhev RHEL7.0.z 4/4] block: add __com.redhat_change-backing-file qmp command
Bugzilla: 1122925
RH-Acked-by: Kevin Wolf <kwolf@redhat.com>
RH-Acked-by: Eric Blake <eblake@redhat.com>
RH-Acked-by: Stefan Hajnoczi <stefanha@redhat.com>

Upstream, the QAPI command 'change-backing-file' was added in a patch,
as part of a series that allowed libvirt to specify backing file
names in block-commit and block-stream.

This standalone QAPI (change-backing-file) is not currently in use by
libvirt, except as a witness that QEMU has the ability to change backing
files in block-commit and block-stream.

However, change-backing-file also relies on node-names functionality,
which is not present in RHEL7.0.  The backport of node-names would also
be fairly instrusive, especially for a command that is not going to be
used beyond verifying its presence.

This downstream patch adds __com.redhat_change-backing-file, which
accepts the same arguments as its upstream namesake.  However, no action
is performed, and the command always returns QERR_UNSUPPORTED.

This is not placed inside the RHEL/RHEV differentiation commands, as it
is not a live block operation (upstream or the dummy version), and
returns synchronously.

Signed-off-by: Jeff Cody <jcody@redhat.com>
Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>