kvmclock: Ensure time in migration never goes backward
Message-id: <20140922170613.897959809@redhat.com>
Patchwork-id: 61360
O-Subject: [PATCH RHEV-7.1 qemu-kvm-rhev 3/3] kvmclock: Ensure time in migration never goes backward
Bugzilla: 1143054
RH-Acked-by: Eduardo Habkost <ehabkost@redhat.com>
RH-Acked-by: Radim Krcmar <rkrcmar@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>

commit 9a48bcd1b82494671c111109b0eefdb882581499 upstream of qemu-kvm.git uq/master branch
Author: Alexander Graf <agraf@suse.de>

When we migrate we ask the kernel about its current belief on what the guest
time would be. However, I've seen cases where the kvmclock guest structure
indicates a time more recent than the kvm returned time.

To make sure we never go backwards, calculate what the guest would have seen as time at the point of migration and use that value instead of the kernel returned one when it's more recent.
This bases the view of the kvmclock after migration on the
same foundation in host as well as guest.

Signed-off-by: Alexander Graf <agraf@suse.de>
Cc: qemu-stable@nongnu.org
Reviewed-by: Marcelo Tosatti <mtosatti@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>