ssh: Enable ssh driver in qemu-kvm-rhev (RHBZ#1138359)
Message-id: <20140905145648.GA27591@redhat.com>
Patchwork-id: 60909
O-Subject: [RHEL-7.1 qemu-kvm-rhev PATCH] ssh: Enable ssh driver in qemu-kvm-rhev (RHBZ#1138359)
Bugzilla: 1138359

virt-v2v has the ability to read from remote RHEL 5 Xen servers.  The
old version used to basically do 'scp', but in the new version we use
the remote source as a backing file, using the qemu ssh driver.  This
avoids copying the whole image before conversion, and can be more
efficient because we use fstrim to reduce what we copy.

I have been testing this on RHEL 7 and QA also have a patched copy of
qemu with this patch which they will use for Xen import testing.

There's no patch to qemu required, just a patch to qemu-kvm.spec as
below.  (Note this is applied on top of the previous patch enabling
https).

Signed-off-by: Richard Jones <rjones@redhat.com>
Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>