simpletrace: install simpletrace.py
Message-id: <1413908245-19510-8-git-send-email-stefanha@redhat.com>
Patchwork-id: 61768
O-Subject: [RHEL7.1 qemu-kvm-rhev PATCH 7/8] simpletrace: install simpletrace.py
Bugzilla: 1155015
RH-Acked-by: Miroslav Rezanina <mrezanin@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>

The simpletrace.py script can be used as a Python module to implement
custom trace analysis scripts.  It can also be used stand-alone to
pretty-print trace files.

Install and ship simpletrace.py so it can be used to analyze SystemTap
simpletrace output.

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>
---
Note this was reimplemented for RHEL7 qemu-kvm because the RHEL6 patch
is not usable due to simpletrace and qemu-kvm.spec.template differences.

Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>