Downstream only: Add script to autoload KVM modules on POWER (v2)
Message-id: <1415160802-25583-1-git-send-email-dgibson@redhat.com>
Patchwork-id: 62131
O-Subject: [RHEV-7.1 qemu-kvm-rhev PATCH] Downstream only: Add script to autoload KVM modules on POWER (v2)
Bugzilla: 1158250
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Gerd Hoffmann <kraxel@redhat.com>

The qemu-kvm-rhev package includes a kvm.modules file which has code to
autoload KVM modules based on platform.  This was inherited from the
Fedora qemu package, but the file isn't actually installed on RHEL or RHEV
at the moment, since it's not needed on x86.

Now that we're going to support POWER RHEV hosts, we want to autoload the
modules.  So, this patch updates the script for current policy and arranges
for it to be installed in the ppc64 package.

Long term it would be nicer for some sort of module alias in the kernel to
trigger an autoload of the KVM module on POWER as on x86, but at present
there isn't really a usable mechanism for this (no POWER equivalent of
the x86cpu alias types).

Signed-off-by: David Gibson <dgibson@redhat.com>
Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>