Downstream only: Remove ISA bus and device support for POWER
Message-id: <1415171263-1587-4-git-send-email-dgibson@redhat.com>
Patchwork-id: 62137
O-Subject: [RHEV-7.1 qemu-kvm-rhev PATCHv3 3/3] Downstream only: Remove ISA bus and device support for POWER
Bugzilla: 1160120
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Gerd Hoffmann <kraxel@redhat.com>
RH-Acked-by: Juan Quintela <quintela@redhat.com>

There's no need for a KVM guest on POWER to have legacy ISA devices, or
even an ISA bus at all.  Since we don't wish to support any extraneous
devices, this removes support for the ISA bridge and all ISA devices from
the configuration for ppc64.

The i8259 interrupt controller is left in, because it can appear in a
lot of contexts and causes other code dependency complications.

Signed-off-by: David Gibson <dgibson@redhat.com>
Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>