hw/arm/virt: Provide flash devices for boot ROMs
Message-id: <1415383153-28089-4-git-send-email-wei@redhat.com>
Patchwork-id: 62222
O-Subject: [RHELSA qemu-kvm-rhev PATCH 03/10] hw/arm/virt: Provide flash devices
        for boot ROMs
Bugzilla: 1160325
RH-Acked-by: Andrew Jones <drjones@redhat.com>

Add two flash devices to the virt board, so that it can be used for
running guests which want a bootrom image such as UEFI. We provide
two flash devices to make it more convenient to provide both a
read-only UEFI image and a read-write place to store guest-set
UEFI config variables. The '-bios' command line option is set up
to provide an image for the first of the two flash devices.

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>
Reviewed-by: Paolo Bonzini <pbonzini@redhat.com>
Message-id: 1409930126-28449-2-git-send-email-ard.biesheuvel@linaro.org

Signed-off-by: Wei Huang <wei@redhat.com>