hw/arm/boot: load device tree to base of DRAM if no -kernel option was passed
Message-id: <1415383153-28089-7-git-send-email-wei@redhat.com>
Patchwork-id: 62216
O-Subject: [RHELSA qemu-kvm-rhev PATCH 06/10] hw/arm/boot: load device tree to
        base of DRAM if no -kernel option was passed
Bugzilla: 1160325
RH-Acked-by: Andrew Jones <drjones@redhat.com>

If we are running the 'virt' machine, we may have a device tree blob but no
kernel to supply it to if no -kernel option was passed. In that case, copy it
to the base of RAM where it can be picked up by a bootloader.

Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Message-id: 1410453915-9344-4-git-send-email-ard.biesheuvel@linaro.org
Signed-off-by: Peter Maydell <peter.maydell@linaro.org>

Signed-off-by: Wei Huang <wei@redhat.com>