hw/arm/virt: mark timer in fdt as v8-compatible
Message-id: <1415383153-28089-9-git-send-email-wei@redhat.com>
Patchwork-id: 62219
O-Subject: [RHELSA qemu-kvm-rhev PATCH 08/10] hw/arm/virt: mark timer in fdt as
        v8-compatible
Bugzilla: 1160325
RH-Acked-by: Andrew Jones <drjones@redhat.com>

check if the first cpu is an armv8 cpu, and if so, put
arm,armv8-timer in the compatible string list.

Note that due to this check, this patch moves the creation
of the timer fdt node to after the cpu creation loop.

Signed-off-by: Claudio Fontana <claudio.fontana@huawei.com>
Message-id: 1411736960-24206-1-git-send-email-hw.claudio@gmail.com
[PMM: updated to list arm,armv8-timer first]
Signed-off-by: Peter Maydell <peter.maydell@linaro.org>

Signed-off-by: Wei Huang <wei@redhat.com>