Downstream only: add script on powerpc to configure CPUs for KVM
Message-id: <1415671257-19875-2-git-send-email-dgibson@redhat.com>
Patchwork-id: 62271
O-Subject: [RHEV 7.1 qemu-kvm-rhev PATCH 2/2] Downstream only: add script on powerpc to configure CPUs for KVM
Bugzilla: 1158250 1158251 1159706
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Gerd Hoffmann <kraxel@redhat.com>

On powerpc we need to configure the CPUs properly in order to run KVM
guests.  Specifically, in addition to loading the KVM module itself, we
need to disable SMT (hardware multithreading) and optionally configure the
number of subcores-per-core (which affects exactly what disabling SMT
does).

This patch adds a script and corresponding systemd unit to the package to
do this necessary setup for KVM.  It's installed only for POWER (although
obviously that could change in future if we need to).

Signed-off-by: David Gibson <dgibson@redhat.com>
Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>