Do not stop qemu-guest-agent service on target switch
Message-id: <1416481847-21227-1-git-send-email-mrezanin@redhat.com>
Patchwork-id: 62487
O-Subject: [RHEL-7.1 qemu-guest-agent PATCH] Do not stop qemu-guest-agent service on target switch
Bugzilla: 1160930
RH-Acked-by: Yash Mankad <ymankad@redhat.com>
RH-Acked-by: Jeff Nelson <jenelson@redhat.com>
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>

From: Miroslav Rezanina <mrezanin@redhat.com>

When systemd target is changed (e.g. change of runlevel using init command),
qemu-guest-agent service is stopped. As it is started as reaction on
creation of virtio device it is not restarted after change.

To prevent stopping of qemu-guest-agent service, we set IgnoreOnIsolate
flag. Howerver, this will cause incorrect reaction on device creation
(i.e. service is not started).

Service has to be changed from static unit to regular one and enabled - so it
is started on device creation. To enable unit after installation, it has to
be put on list of autoenabled units. This is done as part of redhat-release
package so this fix depends on BZ 1165176.

Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1165176

Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>