qemu-iotests: Fix broken test cases
Patchwork-id: 62638
O-Subject: [RHEV-7.1 qemu-kvm-rhev PATCH] qemu-iotests: Fix broken test cases
Bugzilla: 1169589
RH-Acked-by: Jeffrey Cody <jcody@redhat.com>
RH-Acked-by: Max Reitz <mreitz@redhat.com>
RH-Acked-by: Markus Armbruster <armbru@redhat.com>

This fixes a few qemu-iotests cases which currently fail in this tree
because of incomplete or wrong backports.

051 and 087 have trivial output changes.

071 would require whitelisting blkverify, but since this doesn't test
other drivers, but merely the QMP interface for blkdebug/blkverify,
which is only compiled in for debugging and not really supported anyway,
we can just skip the test.

Signed-off-by: Kevin Wolf <kwolf@redhat.com>
Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>

Rebase notes:
  2.2.0: Different context in 087.out

Conflicts:
	tests/qemu-iotests/087.out