Add profiles for Acadia (aarm64) and rhev
This is one of a series of patches to improve git publish. The patches
are designed to support these goals:

a) the default profile should do the right thing;

b) you should be able to specify (the same) profile name in all
   contexts for which it makes sense, even in those contexts where
   that name is the only sensible choice; and

c) there should be a meaningful profile for each kind of development
   work we are doing.

This patch introduces two new profiles for rhev7/master-2.1.2:

1. An Acadia (aarm64) profile. (c)

The Acadia project uses the rhev7/master-2.1.2 branch as its base,
but patches have a different subject than those for rhev, and the
patches are posted to multiple mailing lists.

2. A rhev profile. (b, c)

The rhev7/master-2.1.2 branch is otherwise only for qemu-kvm-rhev.
Provide a profile named 'rhev' that is identical in content to the
profile named 'default'. This means you can use either one of these
commands interchangably:

    git publish --profile rhev
    git publish