Add git publish profiles for rhel and rhev
This is one of a series of patches to improve git publish. The patches
are designed to support these key goals:

a) the default profile should do the right thing;

b) you should be able to specify (the same) profile name in all
   contexts for which it makes sense, even in those contexts where
   that name is the only sensible choice; and

c) there should be a meaningful profile for each kind of development
   work we are doing

This patch introduces two new profiles for rhel70/master:

1. A rhel profile (b, c)

   This is the named profile for development on qemu-kvm in the RHEL-7.0.z
   release.

2. A rhev profile (b, c)

   This is the named profile for development on qemu-kvm-rhev corresponding
   to the RHEL-7.0.z release (schedule).

Note that there is no default profile because I don't believe
a reasonable default exists.