Add git publish profiles for rhel and rhev
RH-Author: Jeff Nelson <jenelson@redhat.com>
Message-id: <55c03be59b217f5efaa738433c096ba67486c28d.1418078036.git.jen@redhat.com>
Patchwork-id: 62774
O-Subject: [RHEL-6.6.z qemu-kvm PATCH] Add git publish profiles for rhel and rhev
Bugzilla:

This is one of a series of patches to improve git publish. The patches
are designed to support these key goals:

a) the default profile should do the right thing;

b) you should be able to specify (the same) profile name in all
   contexts for which it makes sense, even in those contexts where
   that name is the only sensible choice; and

c) there should be a meaningful profile for each kind of development
   work we are doing

This patch introduces two new profiles for rhel70/master:

1. A rhel profile (b, c)

   This is the named profile for development on qemu-kvm in the RHEL-7.0.z
   release.

2. A rhev profile (b, c)

   This is the named profile for development on qemu-kvm-rhev corresponding
   to the RHEL-7.0.z release (schedule).

Note that there is no default profile because I don't believe
a reasonable default exists.

Signed-off-by: Jeff E. Nelson <jen@redhat.com>
---
 .gitpublish | 17 +++++++++++++++++
 1 file changed, 17 insertions(+)
 create mode 100644 .gitpublish

Signed-off-by: Jeff E. Nelson <jen@redhat.com>