Add git publish profiles for rhel and rhev
RH-Author: Jeff Nelson <jenelson@redhat.com>
Message-id: <c4493a319c76079680cea1b0de4c387d47053fe4.1418158445.git.jen@redhat.com>
Patchwork-id: 62804
O-Subject: [RHEL-6.7 qemu-kvm PATCH] Add git publish profiles for rhel and rhev
Bugzilla:

This is one of a series of patches to improve git publish. The patches
are designed to support these key goals:

a) the default profile should do the right thing;

b) you should be able to specify (the same) profile name in all
   contexts for which it makes sense, even in those contexts where
   that name is the only sensible choice; and

c) there should be a meaningful profile for each kind of development
   work we are doing

This patch introduces two new profiles for rhel6/master:

1. A rhel profile (b, c)

This is the named profile for development on qemu-kvm in the RHEL-6.7
release.

2. A rhev profile (b, c)

This is the named profile for development on qemu-kvm-rhev corresponding
to the RHEL-6.7 release (schedule).

Note that there is no default profile because I don't believe
a reasonable default exists.

Signed-off-by: Jeff E. Nelson <jen@redhat.com>
---
 .gitpublish | 17 +++++++++++++++++
 1 file changed, 17 insertions(+)
 create mode 100644 .gitpublish

Signed-off-by: Jeff E. Nelson <jen@redhat.com>